infra/codeexec/src/codeexec/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.28.0
pydantic>=2.7,<3.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0
PyPDF2>=3.0,<4.0
pdfplumber>=0.10,<1.0
reportlab>=4.0,<5.0
python-docx>=1.1,<2.0
google-cloud-storage>=2.13,<3.0
python-multipart>=0.0.6